chore: move builder module into cargo-shuttle
