+++ cargo-shuttle/tests/resources/is-bin/Cargo.toml
