+++ cargo-shuttle/tests/resources/is-bin/src/main.rs
