+++ cargo-shuttle/tests/resources/not-bin/Cargo.toml
