+++ cargo-shuttle/tests/resources/not-bin/src/lib.rs
