+++ cargo-shuttle/tests/resources/not-shuttle/Cargo.toml
