+++ cargo-shuttle/tests/resources/not-shuttle/src/main.rs
