+++ cargo-shuttle/tests/resources/workspace/Cargo.toml
