+++ cargo-shuttle/tests/resources/workspace/alpha2/Cargo.toml
