+++ cargo-shuttle/tests/resources/workspace/alpha2/src/main.rs
