+++ cargo-shuttle/tests/resources/workspace/shared/Cargo.toml
