+++ cargo-shuttle/tests/resources/workspace/shared/src/lib.rs
