--- service/tests/integration/main.rs
@@ -1,2 +1,0 @@
-#[cfg(feature = "builder")]
-mod build_crate;
